Moved AppWindowViewModel and WindowCloser to Switcheroo
