+++ Core/Core.csproj
@@ -46,10 +46,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AppWindow.cs" />
-    <Compile Include="AppWindowViewModel.cs" />
     <Compile Include="IWindowText.cs" />
     <Compile Include="KeyboardHelper.cs" />
-    <Compile Include="WindowCloser.cs" />
     <Compile Include="WindowFilterer.cs" />
     <Compile Include="WindowFinder.cs" />
     <Compile Include="Matchers\ContainsMatcher.cs" />
