+++ Switcheroo/WindowCloser.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Switcheroo.Core
+namespace Switcheroo
 {
 	public class WindowCloser: IDisposable
 	{
